+++ CHANGELOG.md
@@ -35,7 +35,17 @@ Ref: https://keepachangelog.com/en/1.0.0/
 # Changelog
 
 ## [Unreleased]
-- [#668](https://github.com/persistenceOne/pstake-native/pull/668) Update ICA timeout
+
+### Features
+
+- [#667](https://github.com/persistenceOne/pstake-native/pull/667) Monitoring Events.
+
+### Improvements
+
+- [#668](https://github.com/persistenceOne/pstake-native/pull/668) Update ICA timeout.
+
+### Bug Fixes
+
 - [#665](https://github.com/persistenceOne/pstake-native/pull/665) LSM deposit timeout fix.
 
 ## [v2.4.0] - 2023-09-13